:recycle: api: move usecases into enrolment context folder
