+++ api/src/certification/enrolment/application/session-mass-import-controller.js
@@ -1,8 +1,8 @@
 import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import * as csvSerializer from '../../../../lib/infrastructure/serializers/csv/csv-serializer.js';
-import { usecases } from '../../session/domain/usecases/index.js';
-import { getCsvHeaders } from '../infrastructure/files/sessions-import.js';
 import * as csvHelpers from '../../shared/application/helpers/csvHelpers.js';
+import { usecases } from '../domain/usecases/index.js';
+import { getCsvHeaders } from '../infrastructure/files/sessions-import.js';
 
 const createSessions = async function (request, h) {
   const { certificationCenterId } = request.params;
