+++ api/tests/certification/enrolment/integration/domain/usecases/get-mass-import-template-information_test.js
@@ -1,9 +1,9 @@
+import { usecases } from '../../../../../../src/certification/enrolment/domain/usecases/index.js';
 import { CenterTypes } from '../../../../../../src/certification/session/domain/models/CenterTypes.js';
-import { usecases } from '../../../../../../src/certification/session/domain/usecases/index.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect } from '../../../../../test-helper.js';
 
-describe('Integration | Certification | Session | UseCase | get-mass-import-template-information ', function () {
+describe('Integration | Certification | Enrolment | UseCase | get-mass-import-template-information ', function () {
   describe('#getMassImportTemplateInformation', function () {
     it('should return a certification center habilitations and billingMode', async function () {
       // given
